+++ src/main/java/com/controle_de_gastos/notas_api/model/Atributo.java
@@ -13,6 +13,6 @@ import lombok.*;
     public class Atributo {
         @Id
         @GeneratedValue(strategy = GenerationType.IDENTITY)
-        private long idAtributo;
+        private Integer idAtributo;
         private String nome;
     }
